down

+++ docker_scripts/down.sh
@@ -4,4 +4,8 @@ set DIR (dirname (readlink -m (status --current-filename))); cd "$DIR"
 set VENV_PATH ~/.local/robust/$DIR/venv
 . $VENV_PATH/bin/activate.fish ;_
 
-docker-compose  -f ../generated_stack_files/last.yml -p robust --compatibility down
+if test -e ./../generated_stack_files/last.yml;
+	docker-compose  -f ../generated_stack_files/last.yml -p robust --compatibility down
+else;
+	echo ./../generated_stack_files/last.yml not found, nothing to do.
+end
